requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
Markdown==3.6
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2

nltk~=3.8.1
openai~=1.36.0
hanlp_restful~=0.0.23
channels~=3.0.5